Don't use realpath before it's installed on Mac (#87)
The GitHub runner for Mac does not have `realpath` installed. We install
it during the workflow but until we do we shouldn't use it.

In addition I moved `set -euo pipefail` to the top of
`bin/dfx-sns-demo-install`, which would have detected the failing
`realpath`.
And I temporarily disable `set -e` while sourcing `optparse`.