Escape default (#113)
# Motivation
By the way, it seems like clap doesn't allow \< or \> or even spaces in
default values.

Context, I wanted the default to be "dfx-state-backeup-\<generated
timestamp\>". I was then going to check if the default was changed or
not and if not, I would generate a new value. I thought it would look
good in --help like that.

# Changes
* Escape the default variable

# Tests
* A test has been added that exercises a default value with `< >` and
similar characters. It fails without the fix, works with the fix.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>