Test release (#122)
# Motivation
We wish to create snapshot releases automatically. However the releases
don't have to correspond to test runs; tests may do things we are
uninterested in having in snapshots.

# Changes
- Create a script to make a snapshot of a minimal stock state. This may
be used in CI to make Linux snapshots or locally, especially for Macs
that can't use CI snapshots.
- Create a CI job that creates snapshots and, if the current commit is
tagged, adds them to the corresponding releases.
- Create a script taht waits for the replica to create a checkpoint &
use this to make the snapshot after a checkpoint.

# TODO
- Create a workflow to create a release tag; this can be triggered
manually and via cron action.
- Flesh out the stock snapshot so that it contains the canisters and
state we need for CI.

# Tests
- I have pushed tags for dates such as `release-1900-01-01` (with dummy
snapshots for speed) to verify the release logic. The releases can be
viewed here: https://github.com/dfinity/snsdemo/releases They can be
deleted once this testing is complete.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>