Set frontend canister IDs and candid (#133)
# Motivation
In the snsdemo we need to add root as a controller of the frontend
dapps. There is a problem that needs to be fixed first though: The stock
nns frontend dapps are installed by `dfx nns install` (and later updated
to the version we actually want to test). The canister IDs are meant to
be set by `dfx nns import` but the latter doesn't set the canister IDs
for the frontend dapps. It is hard to transfer the ownership of a
canister, the ID of which is not known. Yes, we can hard-wire a canister
ID but this is fragile.

I have a made a PR to fix this upstream. However it typically takes
weeks or more for such a change to trickle down to us. In the meantime,
this PR extends the `dfx-nns-import` polyfill with the missing
functionality.

# Changes
- Set the canister ID and candid for the nns-dapp in `dfx-nns-import`.
- Remove places where the canister IDs were already hard-wired.

# Tests
I have added a test for the `dfx-nns-import` polyfill. It should also be
possible to verify that the upstream fix works as we expect by applying
the same test to `dfx nns import`.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>