Clear dfx wasms cache before 'dfx nns install' (#163)
# Motivation

When running `dfx nns install` you can specify a `DFX_IC_COMMIT` to
determine which versions of canisters will be installed. But if
canisters are already present in `$(dfx cache show)/wasms`, those will
be used regardless of which version they are.
This results in unpredictable canister versions being installed on a
manually created snapshot.
We would like a manually created snapshot to mirror CI as close as
possible to we can test locally under conditions as similar as possible
to CI.

# Changes

1. In `bin/dfx-nns-install`, empty the dfx wasms cache directory before
running `dfx nns install`.
2. Move the removed wasms back to the cache afterwards.

I intentionally didn't remove wasms that weren't there before. This is
debatable but I think it might be useful to have them and it's most
important that the snapshot itself is predictable. Let me know if you
disagree.

# Tests

Ran `bin/dfx-nns-install` (both with empty and non-empty wasms cache)
and watched it empty the cache, download wasms and restore them
afterwards.