Update dfx nns import (#196)
# Motivation
Now that we have dfx v 0.14.1 we have an improved `dfx nns import`. It
is better but not perfect so we need to update the polyfill, we can't
discard it.

# Changes
- Discard the old polyfill
- Use the upstream `dfx nns import` as before, but with support for
networks other than local.
- Replace the internet_identity did file with the one for the release we
are using.

# Notes
## Intermittent test failure on "latest"
The location of the nns lifeline .did file seems to have changed in the
ic repo. With that change, `dfx nns import` breaks. As such, this code
will have to be revised soon.
* The "latest" test failed on Ubutu and Mac, failing to get
https://raw.githubusercontent.com/dfinity/ic/ea6ab1dabffcfb7b901db3ec39002e972ebf69bc/rs/nns/handlers/lifeline/lifeline.did
* Looking in the ic repo, the .did file is now at:
rs/nns/handlers/lifeline/impl/lifeline.did
* However after another push CI is green on "latest". Interesting - has
the location changed back??
## The future of dfx nns import
* `dfx deps` is a new command in dfx for handling canisters that have
already been deployed. It may potentially supersede `dfx nns import`; we
would have to declare the NNS canisters as `"type": "pull"`, provide the
canister IDs and then pull them from mainnet. This functionality is
however unstable so it is probably not a good idea to use it yet.
* `dfx deps` works well for replicating one network in another, it does
not help when installing a specific commit of a project on a testnet. It
looks as if we may need to bring back an updated import command, that
gets remotes and .did definitions from a remote repo.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>