Enable ICRC-2 on ckbtc_ledger (#238)
# Motivation

We want to use ICRC-2 for BTC withdrawal.
The ICRC-1 ledger canister does not yet have ICRC-2 enabled by default.
It's behind a feature flag.

# Change

1. Enable the ICRC-2 feature flag when installing the `ckbtc_ledger`
canister.
2. Drive-by: Reduce the time that the `ckbtc_minter` waits before
sending out BTC. This is 20 minutes by default, for snsdemo it was set
to 7 minutes, but I'm reducing it to 10 seconds. This makes it easier to
verify that a BTC withdrawal actually resulted in a call to the bitcoin
canister.

# Tested

Created a snapshot and ran it to manually make a mock BTC deposit (in
the NNS-dapp UI) and ICRC-2 withdrawal (with `dfx canister call`s).