feat: Installer action for binstall (#264)
# Motivation
We would like to replace `wasm-nm`, which is unmaintained, with
`ic-wasm`. The latter is installed with `binstall` in the `nns-dapp`
repo. We propose to follow the same pattern, the only difference being
where the binstall version is stored.

# Changes
- Add a GitHUb action that installs binstall.#263 

# Tests
The code is exercised heavily in `nns-dapp`, differing only in where
`BINSTALL_VERSION` is sourced from.

Co-authored-by: Max Murphy <max@winning.black>