Create dfx entry if needed (#269)
# Motivation
`dfx nns import` is supposed to create entries for all the NNS canisters
in the local `dfx.json`. Unfortunately with `dfx v0.15.1` the II candid
is not installed and the canister is missing from `dfx.json`. We replace
the candid file with a newer version anyway, that is not a problem, but
we _do_ need the entry in `dfx.json`.

Note: I tested the behaviour of `dfx nns import` for versions `0.14.4`
and `0.15.1` in clean new projects (created with `dfx new
PROJECT_NAME`). This behaviour is in no way specific to the use in the
snsdemo repository.

# Changes
- When updating the II candid, ensure that the path to the candid file
is also listed in `dfx.json`.

# Tests
- For `dfx 0.14.4` see the CI for this PR.
- For `dfx 0.15.1` see the CI for the PR that updates dfx to `0.15.1`.

---------

Co-authored-by: Max Murphy <max@winning.black>