Use 1-proposal in snsdemo (#297)
# Motivation

The 2-proposal flow is deprecated and the 1-proposal flow is simpler.

In snsdemo we have 2 ways of making SNSes:
1. Making a single SNS
2. Making multiple SNSes in parallel

This PR only changes the single SNS flow to 1-proposal. The parallel
flow will be done in another PR.

The 2-proposal flow had a separate step to deploy SNS canisters, which
made it easy to get their canister IDs. The 1-proposal flow does the
deployment automatically so we also need a new script to import those
canister IDs.

**Note**: The first commit copies some files that were used as a base
for 2 new files to make it easier to see what has changed compared to
the originals.

# Changes

1. Add a new `sns_init.yaml` file to use for the 1-proposal flow.
2. Add a flag to `bin/dfx-sns-config-random` to specify which SNS init
template to use, defaults to the old 2-proposal file in order not to
affect the parallel creation flow.
3. Add a new `bin/dfx-sns-propose` script based on the old
`bin/dfx-sns-sale-propose`, with a flag to save the proposal ID to.
4. Add a new `bin/dfx-sns-import-by-proposal` script that imports the
canister IDs created by the 1-proposal flow based on the proposal ID. It
retries many times because the SNSes might not be created immediately.
5. Update `bin/dfx-sns-demo-mksns` to use the 1-proposal flow.

# Tested

1. Created a stock snapshot and ran it from nns-dapp.
2. Checked in the UI that 2 of the 12 SNSes were created with a
1-proposal
3. Checked that I was able to get tokens from the faucet for 1 of them.
4. Checked that I could participate in the swap of the other one.