Do not recreate snsdemo8 identity (#355)
# Motivation

`dfx-sns-demo` deletes and recreates the snsdemo8 identity every time it
is run.
This is unnecessary and can be confusing and inconvenient.

# Changes

1. Instead of deleting the entire snsdemo8 identity, only delete the
neurons file. The neuron will be recreated during the demo and the old
neuron won't exist anymore. But it's expected to get the same neuron ID
because we pass a `--name` which determines the nonce that's used.
2. Only create the snsdemo8 identity if it doesn't exist already.

# Tested

Ran `dfx-sns-demo` with and without pre-existing snsdemo8 identity.