Store selected identity in snapshot (#357)
# Motivation

Currently when starting a snapshot, you need to run `dfx identity use
snsdemo8` before you can upgrade any canisters.
This is because we don't include `identity.json` in the snapshot.

# Changes

Include `identity.json` in the snapshot.

# Tested

Created and started a snapshot. Then ran `dfx identity list` to see
`snsdemo8` selected. The original identity is selected once again when
the snapshot is stopped.