Adjust snsdemo for dfx running pocket-ic (#403)
# Motivation

It appears that with version 0.23.0, `dfx` is no longer running
`icx-proxy` and is now running `pocket-ic`.
I noticed this because trying to update `nns-dapp` to the newest snsdemo
snapshot gave errors: https://github.com/dfinity/nns-dapp/pull/5474
```
+ dfx-sns-propose --network local --save-proposal-id-to /tmp/tmp.6309ZG5i0y
lsof: no process ID specified
lsof 4.93.2
 latest revision: https://github.com/lsof-org/lsof
 latest FAQ: https://github.com/lsof-org/lsof/blob/master/00FAQ
 latest (non-formatted) man page: https://github.com/lsof-org/lsof/blob/master/Lsof.8
 usage: [-?abhKlnNoOPRtUvVX] [+|-c c] [+|-d s] [+D D] [+|-E] [+|-e s] [+|-f[gG]]
 [-F [f]] [-g [s]] [-i [i]] [+|-L [l]] [+m [m]] [+|-M] [-o [o]] [-p s]
 [+|-r [t]] [-s [p:s]] [-S [t]] [-T [t]] [-u s] [+|-w] [-x [fl]] [--] [names]
Use the ``-h'' option to get more help information.
```
I found out that `dfx-sns-propose` needs to get the NNS URL to run
`ic-admin`, for which it uses `dfx-network-provider` which tries to find
the directory that `dfx` was run in via `lsof -p $(pgrep icx-proxy)` but
since `dfx` no longer runs `icx-proxy`, this doesn't work. After
changing `icx-proxy` to `pocket-ic` it seems to work.

There are some other references to `icx-proxy` in the repo which I
decided to fix as well.

# Changes

1. Change `icx-proxy` to `pocket-ic` in `dfx-network-provider`.
2. In several places, replace `pkill icx-proxy` with `dfx killall`.
3. Delete the unused `e2e` directory.
4. Delete the unused scripts `bin/dfx-start-join` and `bin/dfx-status`

# Tested

I was unable to determine in which cases `dfx info replica-port` fails
when not run from the correct directory so when trying to make a test, I
couldn't get the test to fail. I've asked on Slack and may add a test
later.

I did do the following manual test:
1. `dfx start --clean` and `dfx nns install` from snsdemo
2. `ic-admin --nns-url "$(../snsdemo/bin/dfx-network-provider)"
get-node-operator-list` from nns-dapp
This failed without the fix and passed with the fix.