Wait for SNS lifecycle (#409)
# Motivation

SNS swap used to update its lifecycle every second. Now it only does so
once a minute.
When buying into the swap, or when finalizing the swap, it needs to be
in the correct lifecycle.

# Changes

1. Add a script to get the current lifecycle of the current SNS.
2. Add a script to wait for the current SNS to be at a given lifecycle.
3. Use the script to make sure we are at the correct lifecycle before
buying or finalizing.

# Tested

Tried creating a snapshot with the IC commit which has the new swap
canister. It failed without these changes and passes with these (and
other, upcoming) changes.