Use static fake ICP Swap prices instead of random (#438)
# Motivation

In screenshot test in NNS dapp we need to mask out the dollar values
because they differ per snapshot.
This makes testing tedious and reduces covergage.

# Changes

1. Provide static token prices in the fake ICP Swap response.

# Tests

Tested manually. It results in 1 ICP being $9.00, 1 ckBTC being $90'000,
1 ckETH being $3'000, 1 ckUSDC being $1.00 and other tokens being $0.18
each.
While this is somewhat realistic, it's actually not important that these
prices are realistic. Is just nice to have some variety in prices for
testing.